hw03-collections, работа над ошибками, реализоация метода remove(Object)
